+++ backend/core/src/main/java/app/core/model/dto/CreateTransactionBaseRequestDto.java
@@ -7,12 +7,13 @@ import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.Size;
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 //TODO: разделить когда появлятся отличия в income и expense entities
 @Builder
 public record CreateTransactionBaseRequestDto(
-        @NotNull @Positive Long amount,
+        @NotNull @Positive BigDecimal amount,
         @NotNull String category,
         @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ssXXX")
         @NotNull @PastOrPresent OffsetDateTime createDate,
