+++ backend/migration/src/main/resources/db/changelog/1.0.0/create-expenses.sql
@@ -3,7 +3,7 @@ create sequence if not exists expenses_seq;
 create table expenses
 (
     id          bigint not null default nextval('expenses_seq'),
-    amount      bigint,
+    amount      numeric(19,2),
     category    varchar(128),
     create_date timestamp with time zone default now() not null,
     constraint  expenses_pk primary key (id)
